Implement the purchase of tickets
I send SeatIds and ScreeningId to the backend.
It checks to see if everything is OK (e.g., seats have not been sold).
I get a response back to the frontend and display an alert (for now).
I plan to implement an endpoint where you can see
all ticket information and download the PDF.